+++ src/main/java/io/github/mzmine/modules/dataanalysis/significance/anova/AnovaResult.java
@@ -28,7 +28,7 @@ package io.github.mzmine.modules.dataanalysis.significance.anova;
 import io.github.mzmine.datamodel.features.FeatureListRow;
 import io.github.mzmine.modules.dataanalysis.significance.RowSignificanceTestResult;
 
-public record AnovaResult(FeatureListRow row, String groupingColumn, Double pValue) implements
+public record AnovaResult(FeatureListRow row, String groupingColumn, double pValue) implements
     RowSignificanceTestResult {
 
 }
